feat: add the indented line add put function into utils
